Refactoring Web App Creating separate routes for each major function
